Add local address configuration in the HTTP client
Add a configuration option for specifying the local address to bind
sockets to in HTTP clients. The configuration is optional and will, as
before, default to not use a specific address when binding.

Binding to a configured address enables users of the library to, for
example, ensure that a specific network interface is used.