Reset the request content in retry cases before wrapping and buffering it
Content without known length (something that happens with requests for #setBucketLoggingConfiguration() for example)
will get wrapped and buffered, before it eventually ends up as "entity". Marking/Resetting the content stream
returned by BufferedHttpEntity is pretty pointless, as that is *recreated* on every call to #getContent(). At the same
time when the BufferedHttpEntity is created the second time (inside #createHttpRequest()), it's already at its end,
so the buffering produces an *empty* buffer/entity.

This fixes cases where an S3 request with such an entity hits a 307 redirect.