Add possibility to use SSL with a SecurityManager
verifyMasterSecret() reflects on a sun internal, protected package to look at the master secret.
It catches various reflection-related exceptions, but doesn't handle the case where access is denied
by a security policy RuntimePermission("accessClassInPackage.sun.security.ssl"), and doesn't run its
code in a doPrivileged block so that a user can choose to restrict access to this permission just to
the AWS core SDK jar codebase.

With this change users of the API have choices available: to not grant the access at all, or to grant
the access to just the AWS jar. Users not running running with a SecurityManager enabled are not impacted.