MD5ChecksumHandler: Don't advance provided ByteBuffer
Instead of needing to rewind the buffer, the goal is to just ensure that any internal consumptions of the provided ByteBuffers work on read only copies.
This is to ensure that a library user wont be forced to rewind the buffer provided to them (unless they themselves have already consumed the buffer).