Downgrade LOG.errors to LOG.infos on SocketExceptions.
When calling an AWS service, the following can happen:

* SocketException happens in createJsonContent
* LOG.error "Unable to read HTTP response context" is called, and the
SocketException is suppressed
* new JsonContent(null) is called
* new JsonContent constructor calls parseJsonContent
* MAPPER.readTree(null) is called, which causes an NPE
* LOG.error "Unable to parse HTTP response content" is called, and
the NPE is suppressed

The caller of createJsonContent is returned a JsonContent object
with a null node, but this is actually handled correctly (without
any further log.errors), and is turned into an AmazonServiceException:

com.amazonaws.AmazonServiceException: Unable to parse HTTP response content
(Service: AmazonDynamoDBv2; Status Code: 505; Error Code: null; Request ID:
null)

Which is great, as then the request can be retried.

However, the code has issued two LOG.errors for exceptions are suppressed, when
the notification of the failure is going to be passed on to the client as an
AmazonServiceException. E.g. it should be up to the client to call log.error if
this is fatal for them (e.g. after their retries are up).

In our case, the spurious log.errors are causing alerts to happen on
failures that end up being retried and are only transient.