Add "Connection" to ignoredHeaders
When dealing with object metadata, the Connection header may seep into
the metadata map of the ObjectMetadata object, such that when a CopyObjectRequest
is used by cloning the metadata of the source for use at destination, a presence
of the Connection header causes the AWS4Signer to fail as it ends up signing
such a header but the S3 server does not accept this.

Since a Connection header is unnecessary for the purposes of Object Metadata,
it should be ignored by default in the response handler's handling of selective
headers.