Remove unnecessary use of JAXB
In the latest Java 9 ea build (118), the default modules for compile and
runtime were changed (see
http://mail.openjdk.java.net/pipermail/jdk9-dev/2016-May/004309.html).
In particular, java.xml.bind was removed. There is exactly one use of it
here: using DatatypeConverter to encode a base64 string. However, the
Base64Codec used by the Base64 class here already supports encoding. In
version 1.8.3 of the sdk, this was mysteriously changed to use jaxb.

This PR changes the method back to using Base64Codec, as it was before
1230cff. There is no need to use jaxb here, especially since it will
simply not work with java 9 without requiring extra arguments in the
java command line.