Fix overflow problem (#874)
An overflow occured while calculating the expoential backoff.

as MAX_RETRIES is 30, with baseDelay of 5000 you will get to an overflow
at the calculation baseDelay * (1 << retriesAttempted), a negative value will
be returned from Math.min , and passed as the delay value to the call
to Thread.sleep, causing an exception.

My fix simply changes this to (1L << retriesAttempted) , which makes
the multiplication an a long value, avoiding the overflow