aws-java-sdk-core: EC2MetadataClient.java - Adds sending of a User-Agent other than default User-Agent in Java when making requests to metadata service
Similar to AWS Golang SDK and now AWS Ruby SDK, send a User-Agent when making requests to metadata for credentials.
Golang uses a structure like aws-sdk-go/1.8.19 and now Ruby uses aws-sdk-ruby3/ as of PR #1762.  PR #1445 adds this
for python as well.

This will enable protection of AWS credentials from Server Side Request Foregery (SSRF) vectors by use of a metadata
proxy and rejecting User-Agents that do not meet a regex.