Synchronize overridden synchronized methods
All the below overridden methods are marked `synchronized` in the
OpenJDK.  If a synchronized method is overridden in a subclass, and
the overriding method is not synchronized, the thread-safety of the
subclass may be broken.

Since there are no thread-safety comments in these files, I'm not sure
you'll want to merge these.  If we decide not to merge though, it
would be great to document that the classes are no longer thread safe.