Firehose Unable to unmarshall exception content
- Discovered 503 errors did not have json content which caused a
NullPointerException to be thrown and logged consuming our support and
development time. Instead of throwing an exception, we are checking for
null or NullNode before attempting to deserialize the json node to an
exception.
- Additionally the Amazon request header that was being sought was not the
same as the header that is contained in these errors.

See issue: https://github.com/aws/aws-sdk-java/issues/1816