When multipart uploads fail, cancel remaining part transfers
Suppose we have 10 parts, if we fail on part #2 we fire a "TRANFER_FAILED_EVENT" back to the client. We should assume that
clients will react to this failure, possibly with custom logic to retry. Supposing that part #3 succeeds, the client will then get
other transfer events past the "TRANSFER_FAILED_EVENT". I assume here that "TRANSFER_FAILED_EVENT" is a terminal event and that
a client should not expect to receive further events for that transfer, successful or not. However, the meaning of TRANSFER_FAILED_EVENT
is not clearly defined in the javadocs.

fixes: https://github.com/aws/aws-sdk-java/issues/2267