Improve error messages when using proxy server.
If using a proxy server and the AWS exception does not
contain a request ID, then that's a potential sign that the
request was rejected at the proxy server, because the
request ID is assigned at the AWS server side.  This patch
helps users troubleshoot this by including the proxy host in
the exception message.