Match no_proxy environment variable behaviour with other tools
In curl, wget, Ruby, Python, Go and others the no_proxy variable
matches subdomains as well as the parent, as described in this
article:
https://about.gitlab.com/blog/2021/01/27/we-need-to-talk-no-proxy/

The Java nonProxyHosts setting works differently, needing explicit
wildcard characters added as per:
https://docs.oracle.com/javase/8/docs/technotes/guides/net/proxies.html

Especially with path-style URLs for S3 being deprecated it is
helpful to be able to match the behaviour of other tools and libraries
on hosts when trying to set all S3 buckets in a region to bypass
any proxy set.

While this change may impact behaviour, as requests which previously
hit a proxy might now be excluded, the original addition of respecting
the no_proxy environment variable came in 2019 and was likewise a
change in behaviour - as such I believe it is acceptable to make
this change even given the potential impact on clients.