Backported EKS Pod Identity support into ContainerCredentialsProvider
Added support for the AWS_CONTAINER_AUTHORIZATION_TOKEN_FILE environment
variable in the FullUriCredentialsEndpointProvider, which is the
required method of obtaining the authorization token when using the EKS
Pod Identity feature. Moreover, the allowed hosts validation has been
extended to include the ECS and EKS host IPs, which should be allowed
for http (not https) traffic.

Fixes #3062